rename view in academy standart
