+++ js/templates/change-state-game.js
@@ -1,4 +1,4 @@
-import GetStateGame from "./../view/stateGameView";
+import GetStateGame from "../view/state-game-view";
 import {renderTemplate} from "./../utils";
 
 const renderState = (state) => {
