+++ js/view/artist-view.js
