+++ js/view/attempts-ended-view.js
