+++ js/view/result-success-view.js
