+++ js/view/state-game-view.js
