+++ js/view/time-out-view.js
