+++ js/view/welcome-view.js
